Add persistence.xml specifying database connection
There are a few placeholders for anyone that would want to use this.
Specifically, need to update the hibernate.dialect; could be something
like: org.hibernate.dialect.MySQLInnoDBDialect if using MySQL. User
and password are also left as placeholders. In addition, the database
driver for javax.persistence.jdbc.url is set with a placeholder. If
using mysql, it could be: jdbc:mysql://localhost:3306/fantasy_sports