Update DAO classes to work with the database
This commit also sets a default value for createdAt for every
class that has a createdAt column. To verify that database
writes were actually working, I wrote a little test for the
UserDao. It's a crappy test because it doesn't use a separate
test environment database, but that's okay *only for now*.