Add in league-specific player entry and fix class associations
Github issue: #1

As part of trying to build out API endpoints that allow
finding projected and actual scores for a specific player
in a league, I realized that I have to create a league-specific
representation of each player so they can be claimed on many
teams overall, i.e. a player can / and will be on a team in
any fantasy league they exist in.