Implement API endpoint for getting all season scores for a player
Github issue: #1

This commit also restructures the way that the FantasyPlayer
class is used. The intention of the class was to serve as a
way to associate players with fantasy teams, i.e. really is
just a join table in the database. Unfortunately, the
original implementation incorrectly did too much. The class
is renamed as PlayerTeamRecord to make explicit the intention
of this class.